Multiple TIMER and REGEN fixes. Added LASTDISCONNECTED. Fixed LASTUSED.
- Added: LASTDISCONNECTED player keyword, holding the SERV.TIMEHIRES at the time of the last client disconnection.
- Fixed: LASTUSED player attribute was not saved.
- Fixed: Object not being removed from client view when its CONT was changed.
- Fixed: If Timer elapsed on an object in a disconnected or sleeping state, the trigger wouldn't fire when the object got active again (Issues #535, #596).
- Fixed: Timer and REGENs not firing on a ridden characters (Issue #353). Now it fires even if the original character (which gets detatched from its physical REGION) is in a sleeping sector, given that the riding character is in an active (not sleeping nor disconnected) state.

Minor improvements on CServerTime.

+++ src/common/sphereproto.h
@@ -746,7 +746,7 @@ enum RACE_TYPE		// character race, used in new character creation (0x8D) and sta
 #define MAX_ACCOUNT_NAME_SIZE       MAX_NAME_SIZE
 #define MAX_ACCOUNT_PASSWORD_ENTER  16	// client only allows n chars.
 #define ACCOUNT_NAME_VALID_CHAR     " !\"#$%&()*,/:;<=>?@[\\]^{|}~"
-#define MAX_CHARS_PER_ACCT	        7
+#define MAX_CHARS_PER_ACCT	   (byte) 7
 
 #define MAX_SERVERS             32
 #define MAX_SERVER_NAME_SIZE    32
